Restructured components of the cart
